+++ oct/src/vec/vec/test.rs
@@ -8,7 +8,6 @@
 // can obtain one at:
 // <https://mozilla.org/MPL/2.0/>.
 
-use oct::vec;
 use oct::vec::Vec;
 
 #[test]
@@ -33,12 +32,4 @@ fn test_vec_from_iter() {
 		v,
 		[0, 708, 826, 19, 572, 919, 78, 431],
 	);
-}
-
-#[test]
-fn test_vec_macro() {
-	let v0: Vec<u8, 0x4> = vec![0xEF; 0x3];
-	let v1: Vec<u8, 0x4> = vec![0xEF, 0xEF, 0xEF];
-
-	assert_eq!(v0, v1);
 }